Add China UnionPay and Switch and revamp tests
Continued onto the the next cards to see if I could
accomplish a working solution using the standardize
function I had in mind in the first place.  Function
properly executed with the use of forEach and ES6
arrows.  Will have to check again to see the reason
for of loops do not work even when the it statement
is nested within an IIFE. Still stumped why I can't
get by step three given that the logic I utilized
the first time I practiced on test-builder proved to
be sound, while the logic I utilized this time
did not pass the tests. Perhaps the issue lies in
that forEach HoFs prove to be more reliable.